Enable p2 site for DawnVanilla. Add checks that environment variables are set as required. Quote commands passed to ssh.

+++ diamond.releng.jenkins/job.scripts/dawn/dawn_publish_function.sh
@@ -139,11 +139,15 @@ dawn_publish_function () {
     export publish_webserver_diamond_zip_platforms_skipped=0
     if [[ "${publish_webserver_diamond_zip}" == "true" ]]; then
         echo -e "\n*** `date +"%a %d/%b/%Y %H:%M:%S %z"` Publishing .zip to diamond webserver ***\n"
+        if [[ -z "${publish_webserver_diamond_zip_directory_parent}" ]]; then
+            echo "ERROR: $""{publish_webserver_diamond_zip_directory_parent} not set, so terminating"
+            return 100
+        fi
         for platform in linux64 windows64 mac64; do
             platform__indirect="platform_${platform}"
             if [[ "${!platform__indirect}" == "true" ]]; then
                 echo -e "\n*** `date +"%a %d/%b/%Y %H:%M:%S %z"` Publishing to diamond webserver of .zip for ${platform} ***\n"
-                ssh ${webserver_diamond_ssh_options} ${webserver_diamond_name} mkdir -pv ${publish_webserver_diamond_zip_directory_parent}/builds-${name_to_publish_as}/
+                ssh ${webserver_diamond_ssh_options} ${webserver_diamond_name} "mkdir -pv ${publish_webserver_diamond_zip_directory_parent}/builds-${name_to_publish_as}/"
                 ${rsync} -e "${webserver_diamond_rsync_options}" -ilprtDOv --include "/*-${platform}.zip" --exclude '*' ${WORKSPACE}/artifacts_to_publish/. ${webserver_diamond_name}:${publish_webserver_diamond_zip_directory_parent}/builds-${name_to_publish_as}
                 (( publish_webserver_zip_download_platforms_updated += 1 ))
                 # optionally delete any old versions for this platform
@@ -161,11 +165,15 @@ dawn_publish_function () {
     export publish_webserver_opengda_zip_platforms_skipped=0
     if [[ "${publish_webserver_opengda_zip}" == "true" ]]; then
         echo -e "\n*** `date +"%a %d/%b/%Y %H:%M:%S %z"` Publishing .zip to opengda webserver ***\n"
+        if [[ -z "${publish_webserver_opengda_zip_directory_parent}" ]]; then
+            echo "ERROR: $""{publish_webserver_opengda_zip_directory_parent} not set, so terminating"
+            return 100
+        fi
         for platform in linux64 windows64 mac64; do
             platform__indirect="platform_${platform}"
             if [[ "${!platform__indirect}" == "true" ]]; then
                 echo -e "\n*** `date +"%a %d/%b/%Y %H:%M:%S %z"` Publishing to opengda webserver of .zip for ${platform} ***\n"
-                ssh ${webserver_opengda_ssh_options} ${webserver_opengda_name} mkdir -pv ${publish_webserver_opengda_zip_directory_parent}/builds-${name_to_publish_as}/
+                ssh ${webserver_opengda_ssh_options} ${webserver_opengda_name} "mkdir -pv ${publish_webserver_opengda_zip_directory_parent}/builds-${name_to_publish_as}/"
                 ${rsync} -e "${webserver_opengda_rsync_options}" -ilprtDOv --include "/*-${platform}.zip" --exclude '*' ${WORKSPACE}/artifacts_to_publish/. ${webserver_opengda_name}:${publish_webserver_opengda_zip_directory_parent}/builds-${name_to_publish_as}
                 (( publish_webserver_zip_download_platforms_updated += 1 ))
                 # optionally delete any old versions for this platform
@@ -179,13 +187,21 @@ dawn_publish_function () {
     fi
 
     if [ "${publish_p2_site}" == "true" ]; then
+        if [[ -z "${publish_p2_directory_parent}" ]]; then
+            echo "ERROR: $""{publish_p2_directory_parent} not set, so terminating"
+            return 100
+        fi
         temp_suffix=`date +"%Y%m%d_%H%M%S"`
         p2_site_unzipped=${WORKSPACE}/p2_site_unzipped
         rm -rf ${p2_site_unzipped}
         ${unzip} -q ${WORKSPACE}/artifacts_to_publish/*site*.zip -d ${p2_site_unzipped}/
 
         echo -e "\n*** `date +"%a %d/%b/%Y %H:%M:%S %z"` Publishing p2 site ***\n"
-        ssh ${publish_p2_ssh_options} ${publish_p2_server} mkdir -pv ${publish_p2_directory_parent}/
+        if [[ -z "${publish_p2_directory_parent}" ]]; then
+            echo "ERROR: $""{publish_p2_directory_parent} not set, so terminating"
+            return 100
+        fi
+        ssh ${publish_p2_ssh_options} ${publish_p2_server} "mkdir -pv ${publish_p2_directory_parent}/"
         ${rsync} -e "${publish_p2_rsync_options}" -ilprtDOv ${p2_site_unzipped}/ ${publish_p2_server}:${publish_p2_directory_parent}/${name_to_publish_as}_${temp_suffix}/
         # once rsync complete, rename directory
         ssh ${publish_p2_ssh_options} ${publish_p2_server} "rm -rf ${publish_p2_directory_parent}/${name_to_publish_as}/; mv -v ${publish_p2_directory_parent}/${name_to_publish_as}_${temp_suffix}/ ${publish_p2_directory_parent}/${name_to_publish_as}/"
